MessageData log storage encoding
Summary:
Added a 'log storage' serialization format for MessageData, which is now used by the Pilot and Tailer.

* Tenant ID is now 16-bit and encoded in fixed width.
* Re-ordered the normal serialization encoding so that the log storage serialization slice is a sub-slice of the whole serialization. This will allow an optimization where we can avoid doing the log storage serialization again in the Pilot (not implemented).

Test Plan: make check

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D23655

+++ include/Types.h
@@ -108,7 +108,7 @@ typedef std::string URL;
  * tenant can have on the system as a whole. If a tenant exceeds their agreed
  * resource usage then their experience may be throttled.
  */
-typedef uint32_t TenantID;
+typedef uint16_t TenantID;
 
 enum Tenant : TenantID {
   /**
